etl/requirements.txt
pandas>=2.2
pyyaml>=6.0.1
requests>=2.28.0
